+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Use [graceful shutdown](https://trino.io/docs/current/admin/graceful-shutdown.html) for workers ([#461], [#463]).
+- Use [graceful shutdown](https://trino.io/docs/current/admin/graceful-shutdown.html) for workers ([#461], [#463], [#466], [#474]).
 - Default stackableVersion to operator version ([#441]).
 - Configuration overrides for the JVM security properties, such as DNS caching ([#460]).
 
@@ -26,7 +26,9 @@ All notable changes to this project will be documented in this file.
 [#460]: https://github.com/stackabletech/trino-operator/pull/460
 [#461]: https://github.com/stackabletech/trino-operator/pull/461
 [#463]: https://github.com/stackabletech/trino-operator/pull/463
+[#466]: https://github.com/stackabletech/trino-operator/pull/466
 [#470]: https://github.com/stackabletech/trino-operator/pull/470
+[#474]: https://github.com/stackabletech/trino-operator/pull/474
 
 ## [23.7.0] - 2023-07-14